# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "description": "Returns an access token and sets the httpOnly refresh cookie.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed in."
          },
          "401": {
            "description": "Bad credentials, or the account has no password because it uses social sign-in."
          }
        },
        "summary": "Sign in with username and password",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/admin/login": {
      "post": {
        "description": "Username and password only — administrators have no Google or Facebook sign-in, and a social profile can never resolve to an admin account. Accepts `admin` and `super_admin`; a buyer account is rejected. Returns the same access token and refresh cookie as the storefront sign-in.",
        "operationId": "AuthController_adminLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed in as an administrator."
          },
          "401": {
            "description": "Bad credentials, or the account is not an administrator. The message is identical in both cases so this endpoint cannot be used to discover which usernames are privileged."
          }
        },
        "summary": "Sign in as an administrator",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "description": "Always creates a buyer — any `role` sent in the body is ignored, so this cannot be used to self-assign admin. Signs the new user straight in; there is no separate login step and no email verification.",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewUserDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registered and signed in."
          },
          "401": {
            "description": "Username already taken."
          }
        },
        "summary": "Register a new user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "No request body — the httpOnly cookie is the credential. The cookie is rotated on every success; a 401 here is terminal, so send the user to sign in rather than retrying.",
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "New access token issued, refresh cookie rotated."
          },
          "401": {
            "description": "Missing, expired, or already-used refresh token. Reuse revokes the whole session."
          }
        },
        "security": [
          {
            "refresh_token": []
          }
        ],
        "summary": "Rotate the refresh token for a new access token",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "description": "Revokes this token family and clears the cookie. Other devices are unaffected. Returns 204 whether or not a valid cookie was sent.",
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Session ended."
          }
        },
        "security": [
          {
            "refresh_token": []
          }
        ],
        "summary": "End this session",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/google/callback": {
      "get": {
        "description": "Verifies state against the cookie, redeems the code, then redirects to FRONTEND_POST_LOGIN_URL with the refresh cookie set. No token appears in the URL.",
        "operationId": "AuthController_googleCallback",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to the frontend."
          },
          "401": {
            "description": "Missing, mismatched, or wrong-provider state cookie."
          }
        },
        "summary": "Google returns here",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/google/token": {
      "post": {
        "description": "For clients using the Google SDK. Body: { idToken }.",
        "operationId": "AuthController_googleToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderTokenDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed in."
          },
          "400": {
            "description": "idToken missing."
          },
          "401": {
            "description": "ID token failed verification."
          }
        },
        "summary": "Exchange a Google ID token for a session",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/facebook/callback": {
      "get": {
        "operationId": "AuthController_facebookCallback",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to the frontend."
          },
          "401": {
            "description": "Missing, mismatched, or wrong-provider state cookie."
          }
        },
        "summary": "Facebook returns here",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/facebook/token": {
      "post": {
        "description": "For clients using the Facebook SDK. Body: { accessToken } — note Facebook yields an access token, not an ID token.",
        "operationId": "AuthController_facebookToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderTokenDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed in."
          },
          "400": {
            "description": "accessToken missing."
          },
          "401": {
            "description": "Token invalid, or issued for a different Facebook app."
          }
        },
        "summary": "Exchange a Facebook access token for a session",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "description": "Requires a Bearer access token. Returns the profile of whoever the token belongs to — the user id comes from the token, never from the request, so this cannot be used to read another account. The password hash is never included; `hasPassword` reports only whether one is set.",
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponseDTO"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired or invalid token, or the account no longer exists."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get the signed-in user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/all-users": {
      "get": {
        "description": "Requires authentication. Returns a fixed set of safe columns — never the password hash. Has no pagination or role check yet, so it is not suitable for a real admin screen.",
        "operationId": "AuthController_getAllUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List all users",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Clothing Store API",
    "description": "Authentication, catalogue and ordering endpoints. Generated from the controllers, so it cannot drift from the code.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token",
        "description": "httpOnly refresh cookie, scoped to /auth. Set at sign-in and rotated by POST /auth/refresh."
      }
    },
    "schemas": {
      "LoginDTO": {
        "type": "object",
        "properties": {}
      },
      "NewUserDTO": {
        "type": "object",
        "properties": {}
      },
      "ProviderTokenDTO": {
        "type": "object",
        "properties": {}
      },
      "AddressDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 3
          },
          "label": {
            "type": "object",
            "example": "Home",
            "nullable": true
          },
          "line1": {
            "type": "string",
            "example": "221B Galle Road"
          },
          "line2": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "city": {
            "type": "string",
            "example": "Colombo"
          },
          "postalCode": {
            "type": "object",
            "example": "00300",
            "nullable": true
          },
          "country": {
            "type": "string",
            "example": "LK"
          },
          "phone": {
            "type": "object",
            "example": "0771234567",
            "nullable": true
          }
        },
        "required": [
          "id",
          "label",
          "line1",
          "line2",
          "city",
          "postalCode",
          "country",
          "phone"
        ]
      },
      "MeResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 12
          },
          "firstName": {
            "type": "object",
            "example": "Jane",
            "nullable": true
          },
          "lastName": {
            "type": "object",
            "example": "Doe",
            "nullable": true
          },
          "userName": {
            "type": "object",
            "example": "janedoe",
            "nullable": true,
            "description": "Null for accounts created through Google or Facebook."
          },
          "email": {
            "type": "string",
            "example": "jane@example.com"
          },
          "emailVerified": {
            "type": "boolean",
            "example": true
          },
          "phoneNumber": {
            "type": "object",
            "example": "0771234567",
            "nullable": true
          },
          "dob": {
            "type": "object",
            "example": "1995-06-15",
            "nullable": true
          },
          "role": {
            "type": "string",
            "example": "buyer",
            "enum": [
              "buyer",
              "admin",
              "super_admin"
            ]
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-07-01T09:12:44.000Z"
          },
          "hasPassword": {
            "type": "boolean",
            "example": true,
            "description": "False for social-only accounts. Drives whether the profile page offers \"Change password\" or \"Set a password\"."
          },
          "connectedProviders": {
            "example": [
              "google"
            ],
            "description": "Linked identity providers; empty for password-only accounts.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "defaultBillingAddress": {
            "nullable": true,
            "description": "Null until the user saves one.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/AddressDTO"
              }
            ]
          },
          "defaultDeliveryAddress": {
            "nullable": true,
            "description": "Null until the user saves one. May be the same address row as defaultBillingAddress — one address can be flagged as both.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/AddressDTO"
              }
            ]
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "userName",
          "email",
          "emailVerified",
          "phoneNumber",
          "dob",
          "role",
          "isActive",
          "createdAt",
          "hasPassword",
          "connectedProviders",
          "defaultBillingAddress",
          "defaultDeliveryAddress"
        ]
      },
...
    }
  }
}